+++ C++/NumberGuessing/NumberGuessing/Debug/NumberGuessing.log
@@ -1,0 +1,3 @@
+﻿  main.cpp
+  NumberGuessing.vcxproj -> D:\wasse\Documents\GitHub\NumberGuessing\C++\NumberGuessing\Debug\NumberGuessing.exe
+  NumberGuessing.vcxproj -> D:\wasse\Documents\GitHub\NumberGuessing\C++\NumberGuessing\Debug\NumberGuessing.pdb (Full PDB)
