+++ C++/NumberGuessing/NumberGuessing/Debug/NumberGuessing.tlog/NumberGuessing.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=8.1
+Debug|Win32|D:\wasse\Documents\GitHub\NumberGuessing\C++\NumberGuessing\|
