--- NumberGuessing_C++/NumberGuessing/Debug/NumberGuessing.log
@@ -1,3 +1,0 @@
-﻿  NumberGuessing.cpp
-  NumberGuessing.vcxproj -> D:\wasse\Documents\GitHub\NumberGuessing\NumberGuessing_C++\Debug\NumberGuessing.exe
-  NumberGuessing.vcxproj -> D:\wasse\Documents\GitHub\NumberGuessing\NumberGuessing_C++\Debug\NumberGuessing.pdb (Full PDB)
