--- NumberGuessing_C++/NumberGuessing/stdafx.cpp
@@ -1,8 +1,0 @@
-// stdafx.cpp : source file that includes just the standard includes
-// NumberGuessing.pch will be the pre-compiled header
-// stdafx.obj will contain the pre-compiled type information
-
-#include "stdafx.h"
-
-// TODO: reference any additional headers you need in STDAFX.H
-// and not in this file
